Retrieve cards with missing information for Lead Time
If card has completed the cycle (= final startDate is set), but has
missing startDates at the beginning -> information are missing.

This is due because card was already here at the beginning of the period.

We need to retrieve this card specific actions so we can properly
calculate its lead time.