Go with a debounce instead of focusout, for UX
With focusout we need to click somewhere else anytime we change the date,
event using the datepicker. Wich is a bit more annoying.

Debouncing events will give enough time to type a date without perf junk.
Calculation is automatically made after few milliseconds… when we stopped
typing. This seems to be the best alternative.