Fix selectedLabels selected$ merge function
As we did for localStorage, we should also consider that values can
change over time and so newSelected may not be part of it anymore.

Fallback onto props.select() method in such case − which will here deal
with localStorage values… or take the first/last value as the last fallback.