Refactor Trello driver with types
Filter driver inputs regarding an attribute make it more flexible for
different use cases: fetching regular data, fetching missing data…

That will ease the introduction of log in inside the same driver.

Doing so, we are reducing the app complexity. The tradeoff is a format
we should respect for the main outputted observables.