Implement authorization as part of the flow
Now we display a basic welcoming screen to any user. User has to click
on a connect button to trigger Trello authentication process.

Manage rendered DOM and streams regarding the login one = isLogged$.
It will pause other streams until user is connected, so the app can be
kicked-off from the beginning with Cycle.run and it will work!